LayoutTests/imported/w3c/web-platform-tests/cssom/index-002-expected.txt


PASS border is expected to be border: 1px; 
PASS border is expected to be border: 1px solid red; 
PASS border is expected to be border: 1px red; 
PASS border is expected to be border: red; 
PASS border is expected to be border: 1px; (#2) 
PASS border is expected to be border-width: 1px 2px 3px 4px; 
PASS border is expected to be border-width: 2px 1px 1px; 
PASS border is expected to be border-right-width: 1px; border-bottom-width: 1px; border-left-width: 1px; border-top-width: 1px !important; 
PASS border is expected to be border-width: 1px; border-top-color: red; 
PASS border is expected to be border: dotted; 
PASS border is expected to be border-width: 1px; 
FAIL overflow is expected to be overflow: scroll hidden; assert_equals: expected "overflow: scroll hidden;" but got "overflow-x: scroll; overflow-y: hidden;"
PASS overflow is expected to be overflow: scroll; 
PASS outline is expected to be outline: blue dotted 2px; 
PASS margin is expected to be margin: 1px 2px 3px 4px; 
PASS list is expected to be list-style: circle inside; 
PASS list is expected to be list-style-type: lower-alpha; 
PASS font-family is expected to be font-family: sans-serif; line-height: 2em; font-size: 3em; font-style: italic; font-weight: bold; 
PASS padding is expected to be padding: 1px 2px 3px 4px; 
